+++ docs/change_log.rst
@@ -62,6 +62,8 @@ The following are only available within Stem's `git repository
 
  * **Utilities**
 
+  * Added the `stem.util.test_tools <api/util/test_tools.html>`_ module
+  * Started vending the `stem.util.tor_tools <api/util/tor_tools.html>`_ module
   * Added :func:`stem.util.connection.port_usage`
   * Added :func:`stem.util.system.files_with_suffix`